Fix IPv6 handling for cookies (#3564)
